+++ data/tasks.txt
@@ -1,3 +1,3 @@
 T | 0 | borrow book
-D | 0 | finish project | 2024-12-01T16:00:00
-E | 0 | Apply for SEP | 2024-08-22T18:00:00 | 2024-09-22T18:00:00
+D | 0 | finish studying | 2024-12-03T18:00:00
+D | 0 | go to the gym | 2024-10-18T18:00:00
