+++ src/main/java/wolfie/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package wolfie.command;
+
+import wolfie.task.TaskList;
+import wolfie.util.Storage;
+import wolfie.util.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
