+++ src/main/java/wolfie/exception/WolfieException.java
@@ -1,3 +1,5 @@
+package wolfie.exception;
+
 public class WolfieException extends Exception {
     public WolfieException(String message) {
         super(message);
