+++ src/main/java/wolfie/task/Deadline.java
@@ -1,3 +1,5 @@
+package wolfie.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
