+++ src/main/java/wolfie/task/Event.java
@@ -1,3 +1,5 @@
+package wolfie.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
