+++ src/main/java/wolfie/task/TaskType.java
@@ -1,3 +1,5 @@
+package wolfie.task;
+
 public enum TaskType {
     TODO,
     DEADLINE,
