+++ src/main/java/wolfie/task/Todo.java
@@ -1,3 +1,5 @@
+package wolfie.task;
+
 public class Todo extends Task {
     public Todo(String description, boolean isDone) {
         super(description, TaskType.TODO, isDone); // call the constructor of the parent class simple task
