+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello Dean's Lister! I'm Wolfie
+ Hello Dean's Lister! I'm wolfie.Wolfie
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
